(PC-16832) refacto: extract information component in /ui
	and rename it InformationWithIcon
